// src/enums/ColorsEnum.ts
export enum COLORS_ENUM{
  primary = '#48B3C0',
  primary_light = '#7fe5f3', 
  primary_dark = '#008390', 
  primary_text = '#808080',
  primary_warning = '#EDD05A',
  secondary_warning = '#F0E1A3',
  secondary_text = '#C0C0C0',
  background = '#E1E2E1',
  menu_color = '#F5F5F6',
  primary_error = '#ED544E',
  secondary_error = '#F09A97',
  success = '#3AAFA9'
}